Prune dead utility functions and isolate legacy-only helpers
Remove unused functions from active src/utils, including deleting json_utils.py and dead capability helpers.

Create legacy/utils modules for helpers still needed by archived scripts and update legacy imports accordingly.
