+++ legacy/utils/__init__.py
@@ -1,0 +1,1 @@
+"""Utilities retained only for archived legacy workflows."""
