+++ src/utils/capability_management_utils.py
@@ -1,10 +1,8 @@
 """Utility functions for capability management, loading, filtering, and scoring."""
 
-import json
 import logging
 import os
-import random
-from typing import Any, List, Tuple, Union
+from typing import Any, List, Tuple
 
 import torch
 
@@ -16,88 +14,9 @@ from src.utils import constants
 logger = logging.getLogger(__name__)
 
 
-def _sample_seed_capabilities(
-    seed_capability_dir: str,
-    num_seed_capabilities: int = -1,
-    include_capability_names: Union[List[str], None] = None,
-    exclude_capability_names: Union[List[str], None] = None,
-    random_seed: int = 42,
-) -> List[Capability]:
-    """
-    Sample `num_seed_capabilities` seed capabilities from the specified directory.
-
-    These sampled seed capabilities are used in the input prompt
-    to generate new capabilities.
-
-    Args
-    ----
-        seed_capability_dir (str): The directory containing the seed capabilities.
-        num_seed_capabilities (int): The number of seed capabilities to sample.
-        include_capability_names (List[str] | None): A list of
-            capability names to include.
-        exclude_capability_names (List[str] | None): A list of
-            capability names to exclude.
-        random_seed (int): The seed for the random number generator.
-
-    Returns
-    -------
-        List[Capability]: A list of capability objects.
-    """
-    random.seed(random_seed)
-
-    sampled_seed_capabilities = []
-    all_seed_capability_paths = os.listdir(seed_capability_dir)
-
-    if exclude_capability_names is not None:
-        assert num_seed_capabilities != -1, (
-            "Number of seed capabilities should be specified when excluding capabilities."
-        )
-        assert len(exclude_capability_names) < len(all_seed_capability_paths), (
-            "Number of excluded capabilities should be less than the total number of seed capabilities."
-        )
-        assert (
-            len(all_seed_capability_paths) - len(exclude_capability_names)
-        ) >= num_seed_capabilities, (
-            "Number of remaining seed capabilities should be greater than or equal to the number of seed capabilities to sample."
-        )
-        # Remove the excluded capabilities from the list
-        all_seed_capability_paths = [
-            path
-            for path in all_seed_capability_paths
-            if path not in exclude_capability_names
-        ]
-
-    # Select all capabilities if num_seed_capabilities is -1
-    if num_seed_capabilities == -1:
-        num_seed_capabilities = len(all_seed_capability_paths)
-        include_capability_names = None
-
-    # Force include some capabilities
-    if include_capability_names is not None:
-        assert num_seed_capabilities >= len(include_capability_names), (
-            "Number of seed capabilities is less than the number of capabilities to include."
-        )
-        for capability_name in include_capability_names:
-            assert os.path.exists(os.path.join(seed_capability_dir, capability_name)), (
-                f"{capability_name} does not exist in {seed_capability_dir}."
-            )
-            capability = Capability(os.path.join(seed_capability_dir, capability_name))
-            sampled_seed_capabilities.append(capability)
-            all_seed_capability_paths.remove(capability_name)
-        num_seed_capabilities -= len(include_capability_names)
-
-    for capability_path in random.sample(
-        all_seed_capability_paths, num_seed_capabilities
-    ):
-        capability = Capability(os.path.join(seed_capability_dir, capability_path))
-        sampled_seed_capabilities.append(capability)
-
-    return sampled_seed_capabilities
-
-
 def get_previous_capabilities(
     capability_dir: str,
-    capability_area: Union[str, None] = None,
+    capability_area: str | None = None,
     **kwargs: Any,
 ) -> List[Capability]:
     """
@@ -125,27 +44,6 @@ def get_previous_capabilities(
             continue
         prev_capabilities.append(capability)
     return prev_capabilities
-
-
-def get_capability_repr_with_score(capability: Capability, model_name: str) -> str:
-    """
-    Get the capability JSON string with score for the specified model.
-
-    Args
-    ----
-        capability (Capability): The capability to get the JSON string for.
-        model_name (str): The name of the model to use for scoring the capability.
-
-    Returns
-    -------
-        str: A JSON string containing the capability JSON string and score.
-    """
-    if not hasattr(capability, "scores") or model_name not in capability.scores:
-        capability.load_scores(subject_llm_name=model_name)
-    model_score = capability.scores[model_name]
-    capability_dict = capability.to_dict()
-    capability_dict["score"] = model_score
-    return json.dumps(capability_dict, indent=4)
 
 
 def filter_capabilities(
